feat: move FF1 RAM interpretation into versioned profile semantics (#134)
* chore: modernize gradle test setup

* ci: avoid duplicate gradle test run

* ci: consolidate gradle validation workflow

* chore: enable gradle configuration cache

* chore: define emulator architecture backlog

* chore: harden mcp rest bridge payloads

* chore: extract mcp tool catalog

* feat: add agent observation contract

* feat: move FF1 RAM interpretation into versioned profile semantics

Phase rules, position field mappings and landmark anchors now live in
resources/profiles/<id>.json under a "semantics" key, so AgentObservationBuilder
holds no per-game constants. Both bundled profiles carry semantics.

observe() only infers phase, position and location when a profile_id is passed;
the observation reports which profile was applied.

---------

+++ knes-emulator/src/test/kotlin/knes/emulator/ppu/PPURegisterTest.kt
@@ -188,7 +188,7 @@ class PPURegisterTest : FunSpec({
 
         test("bits 5-7 = 0b100: f_color = 4") {
             val h = PpuTestHarness()
-            h.ppu.updateControlReg2(0x80.toInt())  // bit 7 only
+            h.ppu.updateControlReg2(0x80)  // bit 7 only
             val status = h.ppu.statusRegsToInt()
             (status shr 6 and 0x7) shouldBe 4
         }
